Refactor code
* Corrected package name of AssertJ.
