+++ src/test/java/com/woowacourse/javaracingcar/RacingCarUtilTest.java
@@ -2,7 +2,8 @@ package com.woowacourse.javaracingcar;
 
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Java6Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
+
 
 class RacingCarUtilTest {
     @Test
