+++ src/test/java/com/woowacourse/javaracingcar/RandomNumberGeneratorTest.java
@@ -4,7 +4,8 @@ import com.woowacourse.javaracingcar.domain.RandomNumberGenerator;
 import com.woowacourse.javaracingcar.interfaces.NumberGenerator;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
+
 
 public class RandomNumberGeneratorTest {
